Restructured folder; added explanation and comparison to R sections to README.
