+++ local_regression/__init__.py
@@ -1,0 +1,1 @@
+"""Initialize the local_regression package."""
