+++ local_regression/_not_fitted_error.py
@@ -1,0 +1,5 @@
+"""Not fitted error class."""
+
+
+class NotFittedError(ValueError):
+    """Own error class to indicate why the error was raised."""
